+++ osm_fieldwork/odk2csv.py
@@ -46,14 +46,14 @@ def main():
     # parser.add_argument("-o","--outfile", default='tmp.csv', help='The output file for JOSM')
     args = parser.parse_args()
 
-    # if verbose, dump to the termina
+    # if verbose, dump to the terminal
     if args.verbose is not None:
-        log.setLevel(logging.DEBUG)
-        ch = logging.StreamHandler(sys.stdout)
-        ch.setLevel(logging.DEBUG)
-        formatter = logging.Formatter("%(threadName)10s - %(name)s - %(levelname)s - %(message)s")
-        ch.setFormatter(formatter)
-        log.addHandler(ch)
+        logging.basicConfig(
+            level=logging.DEBUG,
+            format=("%(threadName)10s - %(name)s - %(levelname)s - %(message)s"),
+            datefmt="%y-%m-%d %H:%M:%S",
+            stream=sys.stdout,
+        )
 
     xmlfiles = list()
     if args.instance.find("*") >= 0: